adds signin, button, and map components
