+++ mobile/src/components/map/map.js
@@ -1,0 +1,20 @@
+var React = require('react-native');
+
+var {
+  Component,
+  StyleSheet,
+  MapView
+} = React;
+
+module.exports = React.createClass({
+  render: function(){
+    return(
+      <MapView style={styles.map}></MapView>);
+  }
+})
+
+var styles = StyleSheet.create ({
+  map: {
+    flex: 1
+  }
+});
